Extract remaining parts
