+++ core/src/main/scala/akka/persistence/r2dbc/internal/BySliceQuery.scala
@@ -5,6 +5,7 @@
 package akka.persistence.r2dbc.internal
 
 import java.time.Instant
+import java.time.{ Duration => JDuration }
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.Future
@@ -62,7 +63,7 @@ import org.slf4j.Logger
   trait Dao[SerializedRow] {
     def currentDbTimestamp(): Future[Instant]
 
-    def eventsBySlices(
+    def rowsBySlices(
         entityTypeHint: String,
         minSlice: Int,
         maxSlice: Int,
@@ -84,6 +85,11 @@ import org.slf4j.Logger
   import BySliceQuery._
   import TimestampOffset.toTimestampOffset
 
+  private val backtrackingWindow = JDuration.ofMillis(settings.querySettings.backtrackingWindow.toMillis)
+  private val halfBacktrackingWindow = backtrackingWindow.dividedBy(2)
+  private val firstBacktrackingQueryWindow =
+    backtrackingWindow.plus(JDuration.ofMillis(settings.querySettings.backtrackingBehindCurrentTime.toMillis))
+
   def currentBySlices(
       logPrefix: String,
       entityTypeHint: String,
@@ -92,8 +98,8 @@ import org.slf4j.Logger
       offset: Offset): Source[Envelope, NotUsed] = {
     val initialOffset = toTimestampOffset(offset)
 
-    def nextOffset(state: QueryState, eventEnvelope: Envelope): QueryState =
-      state.copy(latest = extractOffset(eventEnvelope), rowCount = state.rowCount + 1)
+    def nextOffset(state: QueryState, envelope: Envelope): QueryState =
+      state.copy(latest = extractOffset(envelope), rowCount = state.rowCount + 1)
 
     def nextQuery(state: QueryState, toDbTimestamp: Instant): (QueryState, Option[Source[Envelope, NotUsed]]) = {
       // FIXME why is this rowCount -1 of expected?, see test EventsBySliceSpec "read in chunks"
@@ -102,7 +108,8 @@ import org.slf4j.Logger
 
         if (state.queryCount != 0 && log.isDebugEnabled())
           log.debug(
-            "currentEventsBySlices query [{}] from slices [{} - {}], from time [{}] to now [{}]. Found [{}] rows in previous query.",
+            "{} query [{}] from slices [{} - {}], from time [{}] to now [{}]. Found [{}] rows in previous query.",
+            logPrefix,
             state.queryCount,
             minSlice,
             maxSlice,
@@ -112,7 +119,7 @@ import org.slf4j.Logger
 
         newState -> Some(
           dao
-            .eventsBySlices(
+            .rowsBySlices(
               entityTypeHint,
               minSlice,
               maxSlice,
@@ -154,6 +161,136 @@ import org.slf4j.Logger
         }
       }
       .mapMaterializedValue(_ => NotUsed)
+  }
+
+  def liveBySlices(
+      logPrefix: String,
+      entityTypeHint: String,
+      minSlice: Int,
+      maxSlice: Int,
+      offset: Offset): Source[Envelope, NotUsed] = {
+    val initialOffset = toTimestampOffset(offset)
+    val someRefreshInterval = Some(settings.querySettings.refreshInterval)
+
+    if (log.isDebugEnabled())
+      log.debug(
+        "Starting {} query from slices [{} - {}], from time [{}].",
+        logPrefix,
+        minSlice,
+        maxSlice,
+        initialOffset.timestamp)
+
+    def nextOffset(state: QueryState, envelope: Envelope): QueryState = {
+      val offset = extractOffset(envelope)
+      if (state.backtracking) {
+        if (offset.timestamp.isBefore(state.latestBacktracking.timestamp))
+          throw new IllegalArgumentException(
+            s"Unexpected offset [$offset] before latestBacktracking [${state.latestBacktracking}].")
+
+        state.copy(latestBacktracking = offset, rowCount = state.rowCount + 1)
+      } else {
+        if (offset.timestamp.isBefore(state.latest.timestamp))
+          throw new IllegalArgumentException(s"Unexpected offset [$offset] before latest [${state.latest}].")
+
+        state.copy(latest = offset, rowCount = state.rowCount + 1)
+      }
+    }
+
+    def delayNextQuery(state: QueryState): Option[FiniteDuration] = {
+      // FIXME verify that this is correct
+      // the same row comes back and is filtered due to how the offset works
+      val delay =
+        if (0 <= state.rowCount && state.rowCount <= 1) someRefreshInterval
+        else None // immediately if there might be more rows to fetch
+
+      // TODO we could have different delays here depending on how many rows that were found.
+      // e.g. a short delay if rowCount is < some threshold
+
+      if (log.isDebugEnabled)
+        delay.foreach { d =>
+          log.debug(
+            "{} query [{}] from slices [{} - {}] delay next [{}] ms.",
+            logPrefix,
+            state.queryCount,
+            minSlice,
+            maxSlice,
+            d.toMillis)
+        }
+
+      delay
+    }
+
+    def nextQuery(state: QueryState): (QueryState, Option[Source[Envelope, NotUsed]]) = {
+      val newIdleCount = if (state.rowCount == 0) state.idleCount + 1 else 0
+      val newState =
+        if (settings.querySettings.backtrackingEnabled && !state.backtracking &&
+          (newIdleCount >= 5 || JDuration
+            .between(state.latestBacktracking.timestamp, state.latest.timestamp)
+            .compareTo(halfBacktrackingWindow) > 0)) {
+          // FIXME config for newIdleCount >= 5 and maybe something like `newIdleCount % 5 == 0`
+
+          // switching to backtracking
+          val fromOffset =
+            if (state.latestBacktracking == TimestampOffset.Zero && initialOffset == TimestampOffset.Zero)
+              TimestampOffset.Zero
+            else if (state.latestBacktracking == TimestampOffset.Zero)
+              TimestampOffset.Zero.copy(timestamp = initialOffset.timestamp.minus(firstBacktrackingQueryWindow))
+            else
+              state.latestBacktracking
+
+          // FIXME the backtracking until offset is state.latest (not equal), but we should probably have an
+          // additional lag to support async (at-least-once) projections without too many duplicates from
+          // backtracking. For exactly-once I think all duplicates are filtered as expected.
+
+          state.copy(
+            rowCount = 0,
+            queryCount = state.queryCount + 1,
+            idleCount = newIdleCount,
+            backtracking = true,
+            latestBacktracking = fromOffset)
+        } else if (state.backtracking && state.rowCount < settings.querySettings.bufferSize - 1) {
+          // switch from backtracking
+          state.copy(rowCount = 0, queryCount = state.queryCount + 1, idleCount = newIdleCount, backtracking = false)
+        } else {
+          state.copy(rowCount = 0, queryCount = state.queryCount + 1, idleCount = newIdleCount)
+        }
+
+      // FIXME for backtracking we could consider to use behind latest offset instead of behind current db time
+      val behindCurrentTime =
+        if (newState.backtracking) settings.querySettings.backtrackingBehindCurrentTime
+        else settings.querySettings.behindCurrentTime
+
+      if (log.isDebugEnabled())
+        log.debug(
+          "{} query [{}]{} from slices [{} - {}], from time [{}]. {}",
+          logPrefix,
+          newState.queryCount,
+          if (newState.backtracking) " in backtracking mode" else "",
+          minSlice,
+          maxSlice,
+          newState.nextQueryFromTimestamp,
+          if (newIdleCount >= 3) s"Idle in [$newIdleCount] queries."
+          else if (state.backtracking) s"Found [${state.rowCount}] rows in previous backtracking query."
+          else s"Found [${state.rowCount}] rows in previous query.")
+
+      newState ->
+      Some(
+        dao
+          .rowsBySlices(
+            entityTypeHint,
+            minSlice,
+            maxSlice,
+            newState.nextQueryFromTimestamp,
+            newState.nextQueryUntilTimestamp,
+            behindCurrentTime)
+          .via(deserializeAndAddOffset(newState.currentOffset)))
+    }
+
+    ContinuousQuery[QueryState, Envelope](
+      initialState = QueryState.empty.copy(latest = initialOffset),
+      updateState = nextOffset,
+      delayNextQuery = delayNextQuery,
+      nextQuery = nextQuery)
   }
 
   // TODO Unit test in isolation
